chore(wallet): remove lib context
