--- components/brave_wallet_ui/common/context/lib.context.ts
@@ -1,8 +1,0 @@
-// Copyright (c) 2022 The Brave Authors. All rights reserved.
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this file,
-// you can obtain one at https://mozilla.org/MPL/2.0/.
-import * as React from 'react'
-import * as Lib from '../async/lib'
-
-export const LibContext = React.createContext(Lib)
