+++ components/brave_wallet_ui/common/hardware/types.ts
@@ -2,10 +2,11 @@
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // you can obtain one at https://mozilla.org/MPL/2.0/.
-import { BraveWallet } from '../../constants/types'
+import { BraveWallet, FilecoinNetwork } from '../../constants/types'
 import { LedgerError } from './ledgerjs/ledger-messages'
 import { EthereumSignedTx } from './ledgerjs/eth-ledger-messages'
 import { FilSignedLotusMessage } from './ledgerjs/fil-ledger-messages'
+import { HardwareVendor } from '../api/hardware_keyrings'
 
 export type HardwareWalletResponseCodeType =
   | 'deviceNotConnected'
@@ -78,3 +79,17 @@ export type GetAccountsHardwareOperationResult = HardwareOperationResult & {
 
 // Batch size of accounts imported from the device in one step.
 export const DerivationBatchSize = 4
+
+export interface HardwareWalletConnectOpts {
+  hardware: HardwareVendor
+  // TODO: add currency and network as enums
+  // currency: string
+  // network: string
+
+  startIndex: number
+  stopIndex: number
+  scheme?: HardwareDerivationScheme
+  network?: FilecoinNetwork
+  coin: BraveWallet.CoinType
+  onAuthorized: () => void
+}
