--- components/brave_wallet_ui/common/hooks/useLib.ts
@@ -1,10 +1,0 @@
-// Copyright (c) 2022 The Brave Authors. All rights reserved.
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this file,
-// you can obtain one at https://mozilla.org/MPL/2.0/.
-import * as React from 'react'
-import { LibContext } from '../context/lib.context'
-
-export const useLib = () => {
-  return React.useContext(LibContext)
-}
