+++ components/brave_wallet_ui/page/screens/send/android/send.tsx
@@ -17,7 +17,6 @@ import 'emptykit.css'
 
 // Utils
 import { loadTimeData } from '../../../../../common/loadTimeData'
-import * as Lib from '../../../../common/async/lib'
 
 // actions
 import * as WalletActions from '../../../../common/actions/wallet_actions'
@@ -29,7 +28,6 @@ import {
   default as BraveCoreThemeProvider
 } from '../../../../../common/BraveCoreThemeProvider'
 import { SendScreen } from '../send_screen/send_screen'
-import { LibContext } from '../../../../common/context/lib.context'
 
 export function AndroidSendApp() {
   return (
@@ -39,13 +37,11 @@ export function AndroidSendApp() {
           dark={walletDarkTheme}
           light={walletLightTheme}
         >
-          <LibContext.Provider value={Lib}>
-            <Switch>
-              <Route>
-                <SendScreen isAndroid={true} />
-              </Route>
-            </Switch>
-          </LibContext.Provider>
+          <Switch>
+            <Route>
+              <SendScreen isAndroid={true} />
+            </Route>
+          </Switch>
         </BraveCoreThemeProvider>
       </BrowserRouter>
     </Provider>
