+++ components/brave_wallet_ui/page/wallet_page.tsx
@@ -14,14 +14,12 @@ import faveiconUrl from '../assets/svg-icons/brave-icon.svg'
 
 // utils
 import { loadTimeData } from '../../common/loadTimeData'
-import * as Lib from '../common/async/lib'
 import { removeDeprecatedLocalStorageKeys } from '../common/constants/local-storage-keys'
 
 // actions
 import * as WalletActions from '../common/actions/wallet_actions'
 
 // contexts
-import { LibContext } from '../common/context/lib.context'
 import { ApiProxyContext } from '../common/context/api-proxy.context'
 
 // components
@@ -37,8 +35,7 @@ import 'emptykit.css'
 import { setIconBasePath } from '@brave/leo/react/icon'
 setIconBasePath('chrome://resources/brave-icons')
 
-function App () {
-
+function App() {
   React.useEffect(() => {
     /** Sets FAVICON for Wallet Pages */
     let link = document.querySelector("link[rel~='icon']") as HTMLLinkElement
@@ -62,9 +59,7 @@ function App () {
           light={walletLightTheme}
         >
           <ApiProxyContext.Provider value={walletPageApiProxy}>
-            <LibContext.Provider value={Lib}>
-              <Container />
-            </LibContext.Provider>
+            <Container />
           </ApiProxyContext.Provider>
         </BraveCoreThemeProvider>
       </BrowserRouter>
